+++ editor/src/components/inspector/sections/style-section/style-section.tsx
@@ -10,8 +10,6 @@ import { TransformSubsection } from './transform-subsection/transform-subsection
 import { Substores, useEditorState } from '../../../editor/store/store-hook'
 import { getInspectorPreferencesForTargets } from '../../../../core/property-controls/property-controls-utils'
 import { when } from '../../../../utils/react-conditionals'
-import { MetadataUtils } from '../../../../core/model/element-metadata-utils'
-import { GridPlacementSubsection } from './container-subsection/grid-cell-subsection'
 
 export enum StyleSubsection {
   Layer = 'Layer',
@@ -43,17 +41,8 @@ export const StyleSection = React.memo(() => {
   const shouldTextContainerSection = inspectorPreferences.includes('typography')
   const shouldShowVisualSections = inspectorPreferences.includes('visual')
 
-  const shouldShowGridCellSection = useEditorState(
-    Substores.metadata,
-    (store) =>
-      store.editor.selectedViews.length === 1 &&
-      MetadataUtils.isGridCell(store.editor.jsxMetadata, store.editor.selectedViews[0]),
-    'StyleSection shouldShowGridCellSection',
-  )
-
   return (
     <React.Fragment>
-      {when(shouldShowGridCellSection, <GridPlacementSubsection />)}
       {when(shouldShowContainerSection, <ContainerSubsection />)}
       {when(shouldTextContainerSection, <TextSubsection />)}
       {when(
